Final Dynamo changes

+++ components/hooks/user_data/use_dynamo_delete.ts
@@ -1,11 +1,8 @@
-import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { QueryKey, useMutation, useQueryClient } from "@tanstack/react-query";
 import { useAppSession } from "../use_app_session";
-import { IModelProps, ModelUuid } from "../../../store/reducers/printing/types";
 import { deleteFromDynamo } from "../../aws/dynamo/dynamo_utilities";
-import { DeleteCommandOutput } from "@aws-sdk/lib-dynamodb";
 import {
-  modelQueryKey,
-  modelsQueryKey,
+  mealPlanQueryKey,
   recipeQueryKey,
   recipesQueryKey,
   shared,
@@ -14,44 +11,76 @@ import {
   IRecipe,
   RecipeUuid,
 } from "../../../store/reducers/food/recipes/types";
+import { IRecipeCacheContext, sharedUpload } from "./use_dynamo_put";
+import { IMealPlan } from "../../../store/reducers/food/meal_plan/types";
+import clone from "just-clone";
 
-export const useDeleteModelFromDynamo = () => {
-  const { id, loading } = useAppSession();
-  const userId = id ?? shared;
+const useDeleteRecipeInCache = () => {
   const queryClient = useQueryClient();
 
   return {
-    ...useMutation(
-      (modelId: ModelUuid) => {
-        if (loading) {
-          throw new Error("User loading");
-        }
-        return deleteFromDynamo({ type: "M-", id: modelId }, userId);
-      },
-      {
-        onSuccess: (data: DeleteCommandOutput, modelId: ModelUuid) => {
-          const mQueryKey = modelQueryKey(modelId, userId);
-          const msQueryKey = modelsQueryKey(userId);
+    mutate: (recipeId: RecipeUuid, userId: string) => {
+      const rQueryKey = recipeQueryKey(recipeId, userId);
+      const rsQueryKey = recipesQueryKey(userId);
 
-          queryClient.removeQueries({ queryKey: mQueryKey });
-          const previousModelsValue: IModelProps[] | undefined =
-            queryClient.getQueryData(msQueryKey);
-          if (previousModelsValue) {
-            queryClient.setQueryData(msQueryKey, () => {
-              return previousModelsValue.filter((model) => model.uuid !== id);
-            });
-          }
-        },
+      const previousRecipeValue: IRecipe | undefined =
+        queryClient.getQueryData(rQueryKey);
+      queryClient.removeQueries({ queryKey: rQueryKey });
+      const previousRecipesValue: IRecipe[] | undefined =
+        queryClient.getQueryData(rsQueryKey);
+
+      if (previousRecipesValue) {
+        queryClient.setQueryData(rsQueryKey, () =>
+          previousRecipesValue.filter((rec) => rec.uuid !== recipeId)
+        );
       }
-    ),
-    disabled: loading,
+
+      const mQueryKey = mealPlanQueryKey(userId);
+      const previousMealPlan: IMealPlan | undefined =
+        queryClient.getQueryData(mQueryKey);
+      if (previousMealPlan) {
+        const newMealPlan = clone(previousMealPlan);
+        for (const date of Object.keys(newMealPlan)) {
+          const dailyPlan = newMealPlan[date];
+          delete dailyPlan[recipeId];
+        }
+        queryClient.setQueryData(mQueryKey, () => newMealPlan);
+      }
+
+      return {
+        rQueryKey,
+        previousRecipeValue,
+        rsQueryKey,
+        previousRecipesValue,
+        mQueryKey,
+        previousMealPlan,
+      };
+    },
+    reset: (
+      context: IRecipeCacheContext & {
+        mQueryKey: QueryKey;
+        previousMealPlan: IMealPlan | undefined;
+      }
+    ) => {
+      const {
+        rQueryKey,
+        previousRecipeValue,
+        rsQueryKey,
+        previousRecipesValue,
+        mQueryKey,
+        previousMealPlan,
+      } = context;
+      queryClient.setQueryData(rQueryKey, previousRecipeValue);
+      queryClient.setQueryData(rsQueryKey, previousRecipesValue);
+      queryClient.setQueryData(mQueryKey, previousMealPlan);
+    },
   };
 };
 
 export const useDeleteRecipeFromDynamo = () => {
   const { id, loading } = useAppSession();
   const userId = id ?? shared;
-  const queryClient = useQueryClient();
+  const { mutate, reset } = useDeleteRecipeInCache();
 
   return {
     ...useMutation(
@@ -59,22 +88,14 @@ export const useDeleteRecipeFromDynamo = () => {
         if (loading) {
           throw new Error("User loading");
         }
+        if (userId === shared) {
+          return Promise.resolve(sharedUpload);
+        }
         return deleteFromDynamo({ type: "R-", id: recipeId }, userId);
       },
       {
-        onSuccess: (data: DeleteCommandOutput, recipeId: RecipeUuid) => {
-          const rQueryKey = recipeQueryKey(recipeId, userId);
-          const rsQueryKey = recipesQueryKey(userId);
-
-          queryClient.removeQueries({ queryKey: rQueryKey });
-          const previousRecipesValue: IRecipe[] | undefined =
-            queryClient.getQueryData(rsQueryKey);
-          if (previousRecipesValue) {
-            queryClient.setQueryData(rsQueryKey, () => {
-              return previousRecipesValue.filter((rec) => rec.uuid !== id);
-            });
-          }
-        },
+        onMutate: (recipeId) => mutate(recipeId, userId),
+        onError: (_, __, context) => context && reset(context),
       }
     ),
     disabled: loading,
